Refactor server structure; add dark CMS logo
