+++ server/controller/contactController.js
@@ -1,0 +1,105 @@
+const Contact = require('../models/contactSchema');
+const validateEmail = require('../utils/validateEmail');
+const transporter = require('../config/nodeMailer');
+
+
+// Contact Controllers
+export const createContact = async (req, res) => {
+  try {
+    const { name, email, message } = req.body;
+
+    if (!name || !email || !message) {
+      return res.status(400).json({ message: 'Please fill in all fields' });
+    }
+
+    if (!validateEmail(email)) {
+      return res.status(400).json({ message: 'Please provide a valid email address' });
+    }
+
+    // Save contact message to database
+    const contact = new Contact({
+      name,
+      email,
+      message
+    });
+
+    await contact.save();
+
+    // Send email notification
+    const mailOptions = {
+      from: process.env.EMAIL_USER,
+      to: 'blustyles0@gmail.com',
+      subject: `New Contact Form Submission from ${name}`,
+      html: `
+        <h2>New Contact Form Submission</h2>
+        <p><strong>Name:</strong> ${name}</p>
+        <p><strong>Email:</strong> ${email}</p>
+        <br>
+        <p><strong>Message:</strong></p>
+        <p>${message}</p>
+      `
+    };
+
+    await transporter.sendMail(mailOptions);
+
+    res.status(201).json({ message: 'Message sent successfully' });
+  } catch (error) {
+    console.error('Contact form error:', error);
+    res.status(500).json({ message: 'Failed to send message' });
+  }
+};
+
+//Get Contacts
+export const getContact = async (req, res) => {
+  try {
+    const { status } = req.query;
+    let query = {};
+    
+    if (status) {
+      query.status = status;
+    }
+    
+    const messages = await Contact.find(query).sort({ createdAt: -1 });
+    res.json(messages);
+  } catch (error) {
+    console.error('Error fetching contact messages:', error);
+    res.status(500).json({ message: 'Failed to fetch messages' });
+  }
+};
+
+// Update contact message
+export const updateContact = async (req, res) => {
+  try {
+    const { status } = req.body;
+    const message = await Contact.findByIdAndUpdate(
+      req.params.id,
+      { status },
+      { new: true }
+    );
+
+    if (!message) {
+      return res.status(404).json({ message: 'Message not found' });
+    }
+
+    res.json(message);
+  } catch (error) {
+    console.error('Error updating contact message:', error);
+    res.status(500).json({ message: 'Failed to update message' });
+  }
+};
+
+// Delete contact
+export const deleteContact = async (req, res) => {
+  try {
+    const message = await Contact.findByIdAndDelete(req.params.id);
+    
+    if (!message) {
+      return res.status(404).json({ message: 'Message not found' });
+    }
+
+    res.json({ message: 'Message deleted successfully' });
+  } catch (error) {
+    console.error('Error deleting contact message:', error);
+    res.status(500).json({ message: 'Failed to delete message' });
+  }
+};
