+++ server/models/bookingSchema.js
@@ -1,0 +1,16 @@
+const bookingSchema = new mongoose.Schema({
+  name: { type: String, required: true },
+  email: { type: String, required: true },
+  phone: { type: String, required: true },
+  date: { type: Date, required: true },
+  time: { type: String, required: true },
+  service: {
+    name: { type: String, required: true },
+    duration: { type: Number, required: true },
+    price: { type: Number, required: true }
+  },
+  status: { type: String, enum: ['pending', 'confirmed', 'cancelled'], default: 'pending' }
+}, { timestamps: true });
+
+const Booking = mongoose.model('Booking', bookingSchema);
+export default Booking;
