+++ server/models/contactSchema.js
@@ -1,0 +1,9 @@
+const contactSchema = new mongoose.Schema({
+  name: { type: String, required: true },
+  email: { type: String, required: true },
+  message: { type: String, required: true },
+  status: { type: String, enum: ['new', 'read', 'responded'], default: 'new' }
+}, { timestamps: true });
+
+const Contact = mongoose.model('Contact', contactSchema);
+export default Contact;
