+++ server/models/userSchema.js
@@ -1,0 +1,10 @@
+const userSchema = new mongoose.Schema({
+  email: { type: String, required: true, unique: true },
+  password: { type: String, required: true },
+  isAdmin: { type: Boolean, default: false },
+  resetPasswordToken: String,
+  resetPasswordExpires: Date
+}, { timestamps: true });
+
+const User = mongoose.model('User', userSchema);
+export default User;
