+++ server/utils/validateEmail.js
@@ -1,0 +1,5 @@
+const validateEmail = (email) => {
+  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  return emailRegex.test(email);
+};
+export default validateEmail;
