+++ server/utils/validatePhone.js
@@ -1,0 +1,5 @@
+const validatePhone = (phone) => {
+  const phoneRegex = /^\(?([0-9]{3})\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$/;
+  return phoneRegex.test(phone);
+};
+export default validatePhone;
